requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
numpy
pandas~=2.1.4
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.24.0.post1

matplotlib~=3.8.2
streamlit
altair<5
plotly